requirements.txt
djangorestframework==3.5.4
django==1.10.5
